# spec/swagger.yaml
swagger: '2.0'
info:
  title: SuggestGrid
  version: 0.1.32-SNAPSHOT
  description: Personalization made Simple
  x-codegen-settings:
    appendContentHeaders: true
    generateAsyncCode: true
    useMethodPrefix: false
    useModelPostfix: false
    useEnumPostfix: false
    useConstructorsForConfig: false
    iOSUseAppInfoPlist: true
    androidUseAppManifest: true
    collectParameters: false
    csharpDefaultNamespace: SuggestGrid
    javaDefaultPackageName: com.suggestgrid
    brandLabel: SuggestGrid
    userAgent: SUGGESTGRID
    projectName: SuggestGrid
    enableAdditionalModelProperties: false
    preserveParameterOrder: true
    enablePHPComposerVersionString: true
schemes:
  - https
securityDefinitions:
  basicAuth:
    type: basic
    description: HTTP Basic Authentication.
security:
  - basicAuth: []
consumes:
  - application/json
produces:
  - application/json
...
definitions:
...
  TypeRequestBody:
    type: object
    description: Options for creating a type.
    properties:
      rating:
        type: string
        description: The rating type of the type. Could be "explicit" or "implicit", where "implicit" is the default.
...
  MessageResponse:
    type: object
    description: Basic response with a message for success responses.
    properties:
      message:
        type: string
        description: Message of the response.
...
  ErrorResponse:
    type: object
    description: Error response.
    properties:
      error_text:
        type: string
        description: Message of the response.
      error_description:
        type: string
        description: Description of the response.
      error_uri:
        type: string
        description: The URI of the error for more details.
...
  LimitExceededErrorResponse:
    type: object
    description: Error response.
    properties:
      error_text:
        type: string
        description: Message of the response.
      error_description:
        type: string
        description: Description of the response.
      error_uri:
        type: string
        description: The URI of the error for more details.
      used:
        type: integer
        format: int64
        description: The quantity used by the account.
      limit:
        type: integer
        format: int64
        description: The limit quantity of the account.
...
  GetTypesResponse:
    type: object
    description: Type names response.
    properties:
      types:
        type: array
        description: The list of type names
        items:
          type: string
  GetTypeResponse:
    type: object
    description: Get type details response.
    properties:
      rating:
        type: string
        description: Rating type of the type that is either implicit or explicit.
...
paths:
...
  /v1/types:
    get:
      tags:
        - type
      operationId: get_all_types
      summary: Gets All Types
      description: Returns all type names in an array named types.
      responses:
        200:
          description: Get types response.
          schema:
            $ref: '#/definitions/GetTypesResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - type
      operationId: delete_all_types
      summary: Deletes All Types
      description: Deletes ALL the types and ALL the actions.
      responses:
        200:
          description: Get types response.
          schema:
            $ref: '#/definitions/MessageResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/types/{type}:
    get:
      tags:
        - type
      operationId: get_type
      summary: Gets Properties of a Type
      description: Returns the options of a type. The response rating parameter.
      parameters:
        - name: type
          in: path
          description: The name of the type to get properties.
          required: true
          type: string
          format: id
      responses:
        200:
          description: Get type response.
          schema:
            $ref: '#/definitions/GetTypeResponse'
        404:
          description: Type does not exists.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - type
      operationId: create_type
      summary: Creates a Type
      description: Creates a new type.
      parameters:
        - name: type
          in: path
          description: The name of the type.
          required: true
          type: string
          format: id
        - name: settings
          in: body
          description: Optional settings for the rating parameter.
          required: false
          schema:
            $ref: '#/definitions/TypeRequestBody'
      responses:
        201:
          description: Type is created.
          schema:
            $ref: '#/definitions/MessageResponse'
        402:
          description: Type limit reached.
          schema:
            $ref: '#/definitions/LimitExceededErrorResponse'
        409:
          description: Type already exists.
          schema:
            $ref: '#/definitions/ErrorResponse'
        422:
          description: Rating type is not `implicit` or `explicit`.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - type
      operationId: delete_type
      summary: Deletes a Type
      description: |
        Warning: Deletes the type with all of its actions and its recommendation model.
      parameters:
        - name: type
          in: path
          description: The name of the type to be deleted.
          required: true
          type: string
          format: id
      responses:
        200:
          description: Type is deleted
          schema:
            $ref: '#/definitions/MessageResponse'
        404:
          description: Action type does not exists.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  # Action Paths
...